requirements.txt
category_encoders==2.3.0
couchbase==4.1.0
numpy==1.21.4
pandas==1.3.4
python-dotenv==0.20.0
scikit_learn==1.0.1
shiv==1.0.0
streamlit==1.14.1
